Reworked the test to cover all supported drivers
